issue-5895: updated silk to bef85a568e4f5dd3e71e9557ec190972514ecb05 (#6503)
### Notes
Updating `contrib/libs/silk` to the most recent upstream commit
(bef85a568e4f5dd3e71e9557ec190972514ecb05) to sync various fixes

Removed obsolete PAGE_SIZE-related patch and added a new small patch for
`__cxxabiv1::__cxa_eh_globals` declaration (in our repo it's already
declared and the declaration is slightly different - it lacks `noexcept`
specifier for one of the methods).

### Issue
https://github.com/ydb-platform/nbs/issues/5895

+++ contrib/libs/silk/src/util/benchmarks/bounded-queue-bench.cpp
@@ -38,7 +38,7 @@ protected:
         {
             for (uint64_t i = 0; i < CAPACITY; ++i)
             {
-                bool b = queue.enqueue(i);
+                bool b = queue.enqueue(static_cast<int>(i));
                 SILK_ASSERT(b);
             }
         }
